Replace DebugStrMap type with a DebugStrTbl type.
The new type essentially encapsulates a []byte data of a .debug_str
section. This way, one can get strings which are parts of other strings.
For example, there could be a full string entry "new_allocator<char>" in
the .debug_str section. While an attribute in the .debug_info can point
to the full string, some other attribute can point to just a part of the
string "allocator<char>". If we only stored a map (the old DebugStrMap)
of full strings, we will not be able to find an entry for part of the
string in the map. In the new type (DebugStrTbl), we read out strings at
the offset on-demand and do not prefetch a map of full strings.

Improved a few errors messages as part of this change. Also, the test
that has been added requires support for a number of additional
attributes. This support has also been added.

A slight reordering of DIE tree construction has also been done to avoid
infinite recursion. This takes care of case wherein a sibling of a DIE
refers to the DIE itself.